# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "File Nest API",
    "version": "1.0.0",
    "description": "RESTful API for file upload, version management, and shareable public links.\n\nResource hierarchy:\n  /files                                    (files)\n    /files/{fileId}\n      /files/{fileId}/versions              (versions of the file)\n        /files/{fileId}/versions/{versionId}\n          /files/{fileId}/versions/{versionId}/content   (binary payload)\n      /files/{fileId}/shares                (share links)\n  /shares/{token}                           (public access via link)"
  },
  "servers": [
    {
      "url": "https://api.filenest.app/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/v1",
      "description": "Development"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/files/{fileId}/versions": {
      "parameters": [
        { "$ref": "#/components/parameters/FileId" }
      ],
      "get": {
        "tags": ["versions"],
        "summary": "List versions of a file",
        "operationId": "listVersions",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PerPage" }
        ],
        "responses": {
          "200": {
            "description": "Version list (newest first)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Version" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "post": {
        "tags": ["versions"],
        "summary": "Upload a new version of an existing file",
        "description": "Creates a new immutable version. The version number is assigned by the server.",
        "operationId": "createVersion",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": { "type": "string", "format": "binary" },
                  "note": {
                    "type": "string",
                    "description": "Free-text note describing the change"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Version created",
            "headers": {
              "Location": { "schema": { "type": "string", "format": "uri" } }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Version" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "FileId": {
        "in": "path",
        "name": "fileId",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
      "Page": {
        "in": "query",
        "name": "page",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "PerPage": {
        "in": "query",
        "name": "perPage",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
...
      "PayloadTooLarge": {
        "description": "File is too large",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      }
    },
    "schemas": {
...
      "Version": {
        "type": "object",
        "required": ["id", "fileId", "versionNumber", "sizeBytes", "createdAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "fileId": { "type": "string", "format": "uuid" },
          "versionNumber": { "type": "integer", "minimum": 1 },
          "note": { "type": ["string", "null"] },
          "mimeType": { "type": "string" },
          "sizeBytes": { "type": "integer", "format": "int64" },
          "checksum": {
            "type": "string",
            "description": "SHA-256 hex digest"
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "createdBy": { "type": "string", "format": "uuid" },
          "_links": {
            "type": "object",
            "properties": {
              "self": { "type": "string", "format": "uri" },
              "file": { "type": "string", "format": "uri" },
              "content": { "type": "string", "format": "uri" }
            }
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 7807 Problem Details",
        "properties": {
          "type": { "type": "string", "format": "uri", "default": "about:blank" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string", "format": "uri" }
        }
      },
...
    }
  }
}